+++ src/main/java/io/coodoo/framework/jpa/control/JpaRevisionEntityListener.java
@@ -2,7 +2,7 @@ package io.coodoo.framework.jpa.control;
 
 import javax.ejb.EJB;
 import javax.ejb.Stateless;
-import javax.persistence.PostPersist;
+import javax.persistence.PrePersist;
 import javax.persistence.PreUpdate;
 
 import io.coodoo.framework.jpa.boundary.entity.AbstractRevisionDatesEntity;
@@ -18,7 +18,7 @@ public class JpaRevisionEntityListener {
     @EJB // @Inject BUG: Inject not possible in EntityListener, FIX available for WildFly 11.0.0.Alpha1, see: https://issues.jboss.org/browse/WFLY-2387
     JpaRevisionService jpaRevisionService;
 
-    @PostPersist
+    @PrePersist
     public void create(AbstractRevisionDatesEntity entity) {
 
         jpaRevisionService.markCreation(entity);